Enable Black on Connexion API folders (#10545)

+++ airflow/api_connexion/endpoints/dag_run_endpoint.py
@@ -23,7 +23,10 @@ from airflow.api_connexion import security
 from airflow.api_connexion.exceptions import AlreadyExists, BadRequest, NotFound
 from airflow.api_connexion.parameters import check_limit, format_datetime, format_parameters
 from airflow.api_connexion.schemas.dag_run_schema import (
-    DAGRunCollection, dagrun_collection_schema, dagrun_schema, dagruns_batch_form_schema,
+    DAGRunCollection,
+    dagrun_collection_schema,
+    dagrun_schema,
+    dagruns_batch_form_schema,
 )
 from airflow.models import DagModel, DagRun
 from airflow.utils.session import provide_session
@@ -36,11 +39,7 @@ def delete_dag_run(dag_id, dag_run_id, session):
     """
     Delete a DAG Run
     """
-    if (
-        session.query(DagRun)
-            .filter(DagRun.dag_id == dag_id, DagRun.run_id == dag_run_id)
-            .delete() == 0
-    ):
+    if session.query(DagRun).filter(DagRun.dag_id == dag_id, DagRun.run_id == dag_run_id).delete() == 0:
         raise NotFound(detail=f"DAGRun with DAG ID: '{dag_id}' and DagRun ID: '{dag_run_id}' not found")
     return NoContent, 204
 
@@ -51,23 +50,24 @@ def get_dag_run(dag_id, dag_run_id, session):
     """
     Get a DAG Run.
     """
-    dag_run = session.query(DagRun).filter(
-        DagRun.dag_id == dag_id, DagRun.run_id == dag_run_id).one_or_none()
+    dag_run = session.query(DagRun).filter(DagRun.dag_id == dag_id, DagRun.run_id == dag_run_id).one_or_none()
     if dag_run is None:
         raise NotFound("DAGRun not found")
     return dagrun_schema.dump(dag_run)
 
 
 @security.requires_authentication
-@format_parameters({
-    'start_date_gte': format_datetime,
-    'start_date_lte': format_datetime,
-    'execution_date_gte': format_datetime,
-    'execution_date_lte': format_datetime,
-    'end_date_gte': format_datetime,
-    'end_date_lte': format_datetime,
-    'limit': check_limit
-})
+@format_parameters(
+    {
+        'start_date_gte': format_datetime,
+        'start_date_lte': format_datetime,
+        'execution_date_gte': format_datetime,
+        'execution_date_lte': format_datetime,
+        'end_date_gte': format_datetime,
+        'end_date_lte': format_datetime,
+        'limit': check_limit,
+    }
+)
 @provide_session
 def get_dag_runs(
     session,
@@ -91,27 +91,52 @@ def get_dag_runs(
     if dag_id != "~":
         query = query.filter(DagRun.dag_id == dag_id)
 
-    dag_run, total_entries = _fetch_dag_runs(query, session, end_date_gte, end_date_lte, execution_date_gte,
-                                             execution_date_lte, start_date_gte, start_date_lte,
-                                             limit, offset)
+    dag_run, total_entries = _fetch_dag_runs(
+        query,
+        session,
+        end_date_gte,
+        end_date_lte,
+        execution_date_gte,
+        execution_date_lte,
+        start_date_gte,
+        start_date_lte,
+        limit,
+        offset,
+    )
 
-    return dagrun_collection_schema.dump(DAGRunCollection(dag_runs=dag_run,
-                                                          total_entries=total_entries))
+    return dagrun_collection_schema.dump(DAGRunCollection(dag_runs=dag_run, total_entries=total_entries))
 
 
-def _fetch_dag_runs(query, session, end_date_gte, end_date_lte,
-                    execution_date_gte, execution_date_lte,
-                    start_date_gte, start_date_lte, limit, offset):
-    query = _apply_date_filters_to_query(query, end_date_gte, end_date_lte, execution_date_gte,
-                                         execution_date_lte, start_date_gte, start_date_lte)
+def _fetch_dag_runs(
+    query,
+    session,
+    end_date_gte,
+    end_date_lte,
+    execution_date_gte,
+    execution_date_lte,
+    start_date_gte,
+    start_date_lte,
+    limit,
+    offset,
+):
+    query = _apply_date_filters_to_query(
+        query,
+        end_date_gte,
+        end_date_lte,
+        execution_date_gte,
+        execution_date_lte,
+        start_date_gte,
+        start_date_lte,
+    )
     # apply offset and limit
     dag_run = query.order_by(DagRun.id).offset(offset).limit(limit).all()
     total_entries = session.query(func.count(DagRun.id)).scalar()
     return dag_run, total_entries
 
 
-def _apply_date_filters_to_query(query, end_date_gte, end_date_lte, execution_date_gte,
-                                 execution_date_lte, start_date_gte, start_date_lte):
+def _apply_date_filters_to_query(
+    query, end_date_gte, end_date_lte, execution_date_gte, execution_date_lte, start_date_gte, start_date_lte
+):
     # filter start date
     if start_date_gte:
         query = query.filter(DagRun.start_date >= start_date_gte)
@@ -147,13 +172,20 @@ def get_dag_runs_batch(session):
     if data["dag_ids"]:
         query = query.filter(DagRun.dag_id.in_(data["dag_ids"]))
 
-    dag_runs, total_entries = _fetch_dag_runs(query, session, data["end_date_gte"], data["end_date_lte"],
-                                              data["execution_date_gte"], data["execution_date_lte"],
-                                              data["start_date_gte"], data["start_date_lte"],
-                                              data["page_limit"], data["page_offset"])
+    dag_runs, total_entries = _fetch_dag_runs(
+        query,
+        session,
+        data["end_date_gte"],
+        data["end_date_lte"],
+        data["execution_date_gte"],
+        data["execution_date_lte"],
+        data["start_date_gte"],
+        data["start_date_lte"],
+        data["page_limit"],
+        data["page_offset"],
+    )
 
-    return dagrun_collection_schema.dump(DAGRunCollection(dag_runs=dag_runs,
-                                                          total_entries=total_entries))
+    return dagrun_collection_schema.dump(DAGRunCollection(dag_runs=dag_runs, total_entries=total_entries))
 
 
 @security.requires_authentication
@@ -167,8 +199,7 @@ def post_dag_run(dag_id, session):
 
     post_body = dagrun_schema.load(request.json, session=session)
     dagrun_instance = (
-        session.query(DagRun).filter(
-            DagRun.dag_id == dag_id, DagRun.run_id == post_body["run_id"]).first()
+        session.query(DagRun).filter(DagRun.dag_id == dag_id, DagRun.run_id == post_body["run_id"]).first()
     )
     if not dagrun_instance:
         dag_run = DagRun(dag_id=dag_id, run_type=DagRunType.MANUAL.value, **post_body)
